+++ src/Controller/FonctionController.php
@@ -15,10 +15,18 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Contracts\Cache\TagAwareCacheInterface;
 use App\Service\DeleteService;
+use Symfony\Bundle\SecurityBundle\Security;
 
 #[Route('/api/fonction')]
 class FonctionController extends AbstractController
 {
+    private $user;
+
+    public function __construct(Security $security)
+    {
+        $this->user = $security->getUser();
+    }
+
     #[Route(name: 'app_fonction', methods: ["GET"])]
     public function getAll(FonctionRepository $fonctionRepository, SerializerInterface $serializer, TagAwareCacheInterface $cache): JsonResponse
     {
@@ -42,8 +50,15 @@ class FonctionController extends AbstractController
     #[Route(name: 'api_fonction_new', methods: ["POST"])]
     public function create(Request $request, SerializerInterface $serializer, EntityManagerInterface $entityManager, TagAwareCacheInterface $cache): JsonResponse
     {
+        if (!$this->user) {
+            return new JsonResponse(['message' => 'User not authenticated'], JsonResponse::HTTP_UNAUTHORIZED);
+        }
+
         $fonction = $serializer->deserialize($request->getContent(), Fonction::class, 'json', []);
-        $fonction->setStatus("on");
+        $fonction->setStatus("on")
+            ->setCreatedBy($this->user->getId())
+            ->setUpdatedBy($this->user->getId())
+        ;
 
         $entityManager->persist($fonction);
         $entityManager->flush();
@@ -57,8 +72,12 @@ class FonctionController extends AbstractController
     #[Route(path: "/{id}", name: 'api_fonction_edit', methods: ["PATCH"])]
     public function update(Fonction $fonction, UrlGeneratorInterface $urlGenerator, Request $request, SerializerInterface $serializer, EntityManagerInterface $entityManager, TagAwareCacheInterface $cache): JsonResponse
     {
+        if (!$this->user) {
+            return new JsonResponse(['message' => 'User not authenticated'], JsonResponse::HTTP_UNAUTHORIZED);
+        }
+
         $updatedFonction = $serializer->deserialize($request->getContent(), Fonction::class, 'json', [AbstractNormalizer::OBJECT_TO_POPULATE => $fonction]);
-        $updatedFonction->setStatus("on");
+        $updatedFonction->setStatus("on")->setUpdatedBy($this->user->getId());
 
         $entityManager->persist($updatedFonction);
         $entityManager->flush();
@@ -72,6 +91,10 @@ class FonctionController extends AbstractController
     #[Route(path: "/{id}", name: 'api_fonction_delete', methods: ["DELETE"])]
      public function delete(Fonction $fonction, Request $request, DeleteService $deleteService): JsonResponse
     {
+        if (!$this->user) {
+            return new JsonResponse(['message' => 'User not authenticated'], JsonResponse::HTTP_UNAUTHORIZED);
+        }
+
         $data = $request->toArray();
         return $deleteService->deleteEntity($fonction, $data, 'fonction');
     }
